add client test #200
